// frontend/intex/src/types/Resident.ts
export interface Resident {
  residentId: number
  caseControlNo: string | null
  internalCode: string | null
  safehouseId: number | null
  caseStatus: string | null
  sex: string | null
  dateOfBirth: string | null
  birthStatus: string | null
  placeOfBirth: string | null
  religion: string | null
  caseCategory: string | null
  subCatOrphaned: boolean | null
  subCatTrafficked: boolean | null
  subCatChildLabor: boolean | null
  subCatPhysicalAbuse: boolean | null
  subCatSexualAbuse: boolean | null
  subCatOsaec: boolean | null
  subCatCicl: boolean | null
  subCatAtRisk: boolean | null
  subCatStreetChild: boolean | null
  subCatChildWithHiv: boolean | null
  isPwd: boolean | null
  pwdType: string | null
  hasSpecialNeeds: boolean | null
  specialNeedsDiagnosis: string | null
  familyIs4ps: boolean | null
  familySoloParent: boolean | null
  familyIndigenous: boolean | null
  familyParentPwd: boolean | null
  familyInformalSettler: boolean | null
  dateOfAdmission: string | null
  ageUponAdmission: string | null
  presentAge: string | null
  lengthOfStay: string | null
  referralSource: string | null
  referringAgencyPerson: string | null
  assignedSocialWorker: string | null
  initialCaseAssessment: string | null
  reintegrationType: string | null
  reintegrationStatus: string | null
  initialRiskLevel: string | null
  currentRiskLevel: string | null
  dateEnrolled: string | null
  dateClosed: string | null
  createdAt: string | null
}